Initial addition of memory_manager class
